+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.9.6 (June 25, 2022)
+- `FEATURE` Added an `index` parameter to `iter` function's callback.
+
 ## 0.9.5 (June 21, 2022)
 - `FEATURE` Added `sector` and `sectorFill` functions to `Render`.
 - `BUGFIX` Fixed the "bounce" of the `myst.ease.backIn` easing function to be closer to that of `myst.ease.backOut`.